LayoutTests/imported/w3c/web-platform-tests/html/dom/historical-expected.txt


FAIL document.applets is always empty assert_array_equals: lengths differ, expected array [] length 0, got object "[object HTMLCollection]" length 1
FAIL HTMLAppletElement is no more assert_equals: expected (undefined) undefined but got (function) function "function HTMLAppletElement() {
    [native code]
}"
FAIL document.all cannot find applet assert_equals: expected (undefined) undefined but got (object) Element node <applet name="war" align="left"></applet>
FAIL document cannot find applet assert_equals: expected (undefined) undefined but got (object) Element node <applet name="war" align="left"></applet>
FAIL window cannot find applet assert_equals: expected (undefined) undefined but got (object) Element node <applet name="war" align="left"></applet>
FAIL applet is not styled assert_equals: expected "none" but got "left"
PASS HTMLTableDataCellElement interface is removed
PASS HTMLTableHeaderCellElement interface is removed
